Move Chunk to separate file
