+++ src/main.rs
@@ -75,6 +75,7 @@ mod errors;
 mod app;
 mod assets;
 mod backend;
+mod chunk;
 mod clipboard;
 mod color;
 mod config;
